Fix layout for recipe details list item
The original layout had a circular constraint dependency where the
the recipe name depended on the servings TextView to measure its
width and the servings depended on the name TextView to set its
baseline. Also fixed the left margin of the name to utilize the
correct value for the key line.