Fix ExoPlayer media handling
Previously, the media would be loaded to any Fragments created,
regardless of whether they were visible to the user. Additionally,
a new Step and new SimpleExoPlayer would be created every time a new
Fragment was created regardless of whether the savedInstanceState was
recycled by StepSectionAdapter. This caused a memory leak where every
time the user scrolled to a new page, it would create another player.
Additionally, when scrolling back to a tab where the player had
previously been loaded, there would be no control over the media
player the SimpleExoPlayerView did not set its player to the newly
created SimpleExoPlayer. The setPlayer() method was changed to only
create a new SimpleExoPlayer if the member variable was null, but the
MediaSource must be re-created every time. Not sure why, but the
player will not play if the MediaSource is assumed to be attached to
the SimpleExoPlayer.

Additionally, added stopPlayer() method to save the video's position
so it can be set correctly when the SimpleExoPlayer re-prepares the
MediaSource.