Move PageChangeCallBack instances to to StepDetailsAdapter
Reduces the number of CallBacks required as there is only ever one
StepDetailsAdapter. Also fixes an issue where the CallBack would
attempt to would send a message to a Handler on a dead thread. Also,
the Steps created by StepSectionAdapter are saved in an Array so
their respective SimpleExoPlayers can be accessed in onPageChanged().