Refactor video assets out of Step Object
Reduces the number of SimpleExoPlayers to one because it is now bound
to the Activity holding the StepDetailsFragment. Consolidates the
MediaSources to an outside source so they do not need to be created
and held by the Step Object. This also prevents the ExoPlayerView
from playing the video when it binds to the data.