Implement a GlideModule using the Glide annotation processor
The default Glide API uses a signature based on the memory resource
address which causes the EngineKey to change every time the app
restarts. This means that cached images can't be loaded properly and
must be reloaded each time the app is closed from memory.

Added a RecipeGlideSignature class that returns a constant signature to
work around the issue.