Refactor method for retrieving the SimpleExoPlayer to new Interface
Because the SimpleExoPlayer is always bound to an Activity, retrieving
the SimpleExoPlayer was moved to a new Interface to be implemented by
the Activities containing the player. This consolidates the method for
obtaining the player to simply check to ensure the Activity
implements the Interface prior to obtaining the SimpleExoPlayer.