Alter RecipeDetailsActivity to swap the Step instead of entire Fragment
Because the height of the PlayerView in StepDetailsFragment needs to
be set dynamically after its width has been measured, there is a
slight delay and flickering as the height is set each time a
StepDetailsFragment is loaded. Created a new method to just swap the
Step being used bind data to the View to avoid laying out a new
PlayerView every time the user selects a new Step.