Resolve issue with video player position not being properly saved (#1)
Added a check to ensure that the Step's player is valid prior to saving
the video player's position because the Activity would revert
mCurrentPosition to a different position before calling
onSaveInstanceState(), causing a Step without a player to attempt to
save its position. This allows only the Step with a valid video player
to save its position.